requirements.txt
autopep8==1.4.3
awscli==1.16.91
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
botocore==1.12.81
certifi==2019.3.9
chardet==3.0.4
colorama==0.3.9
configparser==3.7.1
docutils==0.14
enum34==1.1.6
future==0.17.1
futures==3.2.0
idna==2.8
jedi==0.13.2
jmespath==0.9.3
lxml==4.3.2
mccabe==0.6.1
mercurial==4.7.2
parso==0.3.4
Pillow==5.4.1
pipenv==2018.11.26
pluggy==0.8.1
pyasn1==0.4.5
pycodestyle==2.5.0
pydocstyle==3.0.0
pyflakes==2.1.0
python-dateutil==2.7.5
python-jsonrpc-server==0.1.2
python-language-server==0.24.0
PyYAML==3.13
requests==2.21.0
rope==0.12.0
rsa==3.4.2
s3transfer==0.1.13
six==1.12.0
snowballstemmer==1.2.1
soupsieve==1.8
typing==3.6.6
urllib3==1.24.1
virtualenv==16.4.3
virtualenv-clone==0.5.1
yapf==0.26.0
